Improved Class and Template selection process
Selecting a Template now makes sense, and dynamically changes as templates are added to the base class.
Reworked the entire indexing process for both the Classes and Templates files.
Templates with modifications to Reactor Stress or Structure now have different effects (adding or setting) the base class statistics.
Classes in Classes.json now have the basic Structure and Reactor Stress stats as a default.
Renamed all of the functions to long form.